original/requirements.txt
anytree
pynacl
ddt